pyproject.toml
[tool.poetry]
name = "fastapi app boilerplate"
version = "0.1.0"
description = "Aplicação modelo para RestAPI usando FastAPI"
authors = ["Marlos Ribeiro <marlos.mgrb@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = { extras = ["email"], version = "^2.7.0" }
pydantic-settings = "^2.2.1"
rich = "^13.7.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "^4.1.3"
requests = "^2.32.3"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.5"
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.17"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
httpx = "^0.27.0"
commitizen = "^3.22.0"
pre-commit = "^3.7.0"
cz-conventional-gitmoji = "^0.2.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
pre_test_all = 'task lint'
test_all = 'pytest -s -x --cov=src -vv'
post_test_all = 'coverage html'
